feat: Chat 컴포넌트 useScrollToBottom 커스텀 훅 모듈화
